Consolidate audio tests and update runtime
